Add duplicate check for tasks
Task can be added without checking for duplicate
entries.

Let's add a duplicate check to ensure that user do
not accidentally add the same task more than once.

Having duplicate task in the list could create
confusion for the user and deleting having to
delete them later is a hassle.